requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.14
openai==1.35.3
chromadb==0.5.3
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
Jinja2==3.1.4
pydantic==2.7.4
